infra: Remove the TCP API
Deprecated in favor of GraphQL
