+++ src/infra/mod.rs
@@ -8,6 +8,5 @@ pub mod ldap_handler;
 pub mod ldap_server;
 pub mod logging;
 pub mod sql_backend_handler;
-pub mod tcp_api;
 pub mod tcp_backend_handler;
 pub mod tcp_server;
